メッセージ送信スレッドを別ファイルに分割
